+++ packages/atomic/src/components/search/atomic-no-results/atomic-no-results.stories.tsx
@@ -1,7 +1,6 @@
 import defaultStory from 'atomic-storybook/default-story';
 
 const {defaultModuleExport, exportedStory} = defaultStory(
-  'Atomic/NoResults',
   'atomic-no-results',
   {},
   {