+++ packages/headless/src/api/commerce/commerce-api-params.ts
@@ -1,4 +1,4 @@
-import {AnyCommerceFacetRequest} from '../../features/commerce/facets/facet-set/interfaces/request';
+import {AnyFacetRequest} from '../../features/commerce/facets/facet-set/interfaces/request';
 import {SortOption} from './common/sort';
 
 export interface TrackingIdParam {
@@ -62,11 +62,15 @@ export interface CartItemParam {
 }
 
 export interface FacetsParam {
-  facets?: AnyCommerceFacetRequest[];
+  facets?: AnyFacetRequest[];
 }
 
 export interface PageParam {
   page?: number;
+}
+
+export interface PerPageParam {
+  perPage?: number;
 }
 
 export interface SortParam {
@@ -83,4 +87,8 @@ export interface FacetQueryParam {
 
 export interface FacetIdParam {
   facetId: string;
+}
+
+export interface SlotIdParam {
+  id: string;
 }