+++ packages/headless/src/features/generated-answer/generated-answer-insight-analytics-actions.ts
@@ -1,8 +1,9 @@
 import {Qna} from '@coveo/relay-event-types';
 import {
   InsightAction,
-  makeInsightAnalyticsAction,
+  makeInsightAnalyticsActionFactory,
 } from '../analytics/analytics-utils';
+import {SearchPageEvents} from '../analytics/search-action-cause';
 import {getCaseContextAnalyticsMetadata} from '../case-context/case-context-state';
 import {
   citationSourceSelector,
@@ -15,10 +16,11 @@ export type GeneratedAnswerFeedback =
   | 'notAccurate'
   | 'outOfDate'
   | 'harmful';
+const RGAType = 'RGA';
 
 //TODO: SFINT-5435
 export const logRetryGeneratedAnswer = (): InsightAction =>
-  makeInsightAnalyticsAction(
+  makeInsightAnalyticsActionFactory(SearchPageEvents.retryGeneratedAnswer)(
     'analytics/generatedAnswer/retry',
     (client, state) =>
       client.logRetryGeneratedAnswer(
@@ -30,7 +32,7 @@ export const logRetryGeneratedAnswer = (): InsightAction =>
 export const logRephraseGeneratedAnswer = (
   responseFormat: GeneratedResponseFormat
 ): InsightAction =>
-  makeInsightAnalyticsAction(
+  makeInsightAnalyticsActionFactory(SearchPageEvents.rephraseGeneratedAnswer)(
     'analytics/generatedAnswer/rephrase',
     (client, state) => {
       const generativeQuestionAnsweringId =
@@ -51,46 +53,51 @@ export const logRephraseGeneratedAnswer = (
 export const logOpenGeneratedAnswerSource = (
   citationId: string
 ): InsightAction =>
-  makeInsightAnalyticsAction({
-    prefix: 'analytics/generatedAnswer/openAnswerSource',
-    __legacy__getBuilder: (client, state) => {
-      const generativeQuestionAnsweringId =
-        generativeQuestionAnsweringIdSelector(state);
-      const citation = citationSourceSelector(state, citationId);
+  makeInsightAnalyticsActionFactory(SearchPageEvents.openGeneratedAnswerSource)(
+    {
+      prefix: 'analytics/generatedAnswer/openAnswerSource',
+      __legacy__getBuilder: (client, state) => {
+        const generativeQuestionAnsweringId =
+          generativeQuestionAnsweringIdSelector(state);
+        const citation = citationSourceSelector(state, citationId);
 
-      if (!generativeQuestionAnsweringId || !citation) {
-        return null;
-      }
-      return client.logOpenGeneratedAnswerSource(
-        {
-          generativeQuestionAnsweringId,
-          permanentId: citation.permanentid,
-          citationId: citation.id,
-        },
-        getCaseContextAnalyticsMetadata(state.insightCaseContext)
-      );
-    },
-    analyticsType: 'Qna.CitationClick',
-    analyticsPayloadBuilder: (state): Qna.CitationClick => {
-      const generativeQuestionAnsweringId =
-        generativeQuestionAnsweringIdSelector(state);
-      return {
-        answer: {
-          id: generativeQuestionAnsweringId!,
-          type: 'CRGA',
-        },
-        citation: {
-          id: citationId,
-        },
-      };
-    },
-  });
+        if (!generativeQuestionAnsweringId || !citation) {
+          return null;
+        }
+        return client.logOpenGeneratedAnswerSource(
+          {
+            generativeQuestionAnsweringId,
+            permanentId: citation.permanentid,
+            citationId: citation.id,
+          },
+          getCaseContextAnalyticsMetadata(state.insightCaseContext)
+        );
+      },
+      analyticsType: 'Qna.CitationClick',
+      analyticsPayloadBuilder: (state): Qna.CitationClick => {
+        const generativeQuestionAnsweringId =
+          generativeQuestionAnsweringIdSelector(state);
+        return {
+          answer: {
+            id: generativeQuestionAnsweringId!,
+            type: RGAType,
+          },
+          citation: {
+            id: citationId,
+            type: 'Source',
+          },
+        };
+      },
+    }
+  );
 
 export const logHoverCitation = (
   citationId: string,
   citationHoverTimeInMs: number
 ): InsightAction =>
-  makeInsightAnalyticsAction({
+  makeInsightAnalyticsActionFactory(
+    SearchPageEvents.generatedAnswerSourceHover
+  )({
     prefix: 'analytics/generatedAnswer/hoverCitation',
     __legacy__getBuilder: (client, state) => {
       const generativeQuestionAnsweringId =
@@ -117,10 +124,11 @@ export const logHoverCitation = (
       return {
         answer: {
           id: generativeQuestionAnsweringId!,
-          type: 'CRGA',
+          type: RGAType,
         },
         citation: {
           id: citationId,
+          type: 'Source',
         },
         citationHoverTimeInMs,
       };
@@ -128,7 +136,7 @@ export const logHoverCitation = (
   });
 
 export const logLikeGeneratedAnswer = (): InsightAction =>
-  makeInsightAnalyticsAction({
+  makeInsightAnalyticsActionFactory(SearchPageEvents.likeGeneratedAnswer)({
     prefix: 'analytics/generatedAnswer/like',
     __legacy__getBuilder: (client, state) => {
       const generativeQuestionAnsweringId =
@@ -150,7 +158,7 @@ export const logLikeGeneratedAnswer = (): InsightAction =>
       return {
         answer: {
           id: generativeQuestionAnsweringId!,
-          type: 'CRGA',
+          type: RGAType,
         },
         feedback: {
           liked: true,
@@ -160,7 +168,7 @@ export const logLikeGeneratedAnswer = (): InsightAction =>
   });
 
 export const logDislikeGeneratedAnswer = (): InsightAction =>
-  makeInsightAnalyticsAction({
+  makeInsightAnalyticsActionFactory(SearchPageEvents.dislikeGeneratedAnswer)({
     prefix: 'analytics/generatedAnswer/dislike',
     __legacy__getBuilder: (client, state) => {
       const generativeQuestionAnsweringId =
@@ -182,7 +190,7 @@ export const logDislikeGeneratedAnswer = (): InsightAction =>
       return {
         answer: {
           id: generativeQuestionAnsweringId!,
-          type: 'CRGA',
+          type: RGAType,
         },
         feedback: {
           liked: false,
@@ -194,7 +202,9 @@ export const logDislikeGeneratedAnswer = (): InsightAction =>
 export const logGeneratedAnswerFeedback = (
   feedback: GeneratedAnswerFeedback
 ): InsightAction =>
-  makeInsightAnalyticsAction({
+  makeInsightAnalyticsActionFactory(
+    SearchPageEvents.generatedAnswerFeedbackSubmit
+  )({
     prefix: 'analytics/generatedAnswer/sendFeedback',
     __legacy__getBuilder: (client, state) => {
       const generativeQuestionAnsweringId =
@@ -217,7 +227,7 @@ export const logGeneratedAnswerFeedback = (
       return {
         answer: {
           id: generativeQuestionAnsweringId!,
-          type: 'CRGA',
+          type: RGAType,
         },
         feedback: {
           liked: false,
@@ -230,7 +240,9 @@ export const logGeneratedAnswerFeedback = (
 export const logGeneratedAnswerDetailedFeedback = (
   details: string
 ): InsightAction =>
-  makeInsightAnalyticsAction({
+  makeInsightAnalyticsActionFactory(
+    SearchPageEvents.generatedAnswerFeedbackSubmit
+  )({
     prefix: 'analytics/generatedAnswer/sendFeedback',
     __legacy__getBuilder: (client, state) => {
       const generativeQuestionAnsweringId =
@@ -254,7 +266,7 @@ export const logGeneratedAnswerDetailedFeedback = (
       return {
         answer: {
           id: generativeQuestionAnsweringId!,
-          type: 'CRGA',
+          type: RGAType,
         },
         feedback: {
           liked: false,
@@ -269,7 +281,7 @@ export const logGeneratedAnswerDetailedFeedback = (
 export const logGeneratedAnswerStreamEnd = (
   answerGenerated: boolean
 ): InsightAction =>
-  makeInsightAnalyticsAction(
+  makeInsightAnalyticsActionFactory(SearchPageEvents.generatedAnswerStreamEnd)(
     'analytics/generatedAnswer/streamEnd',
     (client, state) => {
       const generativeQuestionAnsweringId =
@@ -288,7 +300,9 @@ export const logGeneratedAnswerStreamEnd = (
   );
 
 export const logGeneratedAnswerShowAnswers = (): InsightAction =>
-  makeInsightAnalyticsAction({
+  makeInsightAnalyticsActionFactory(
+    SearchPageEvents.generatedAnswerShowAnswers
+  )({
     prefix: 'analytics/generatedAnswer/show',
     __legacy__getBuilder: (client, state) => {
       const generativeQuestionAnsweringId =
@@ -311,14 +325,16 @@ export const logGeneratedAnswerShowAnswers = (): InsightAction =>
         action: 'show',
         answer: {
           id: generativeQuestionAnsweringId!,
-          type: 'CRGA',
+          type: RGAType,
         },
       };
     },
   });
 
 export const logGeneratedAnswerHideAnswers = (): InsightAction =>
-  makeInsightAnalyticsAction({
+  makeInsightAnalyticsActionFactory(
+    SearchPageEvents.generatedAnswerHideAnswers
+  )({
     prefix: 'analytics/generatedAnswer/hide',
     __legacy__getBuilder: (client, state) => {
       const generativeQuestionAnsweringId =
@@ -341,14 +357,16 @@ export const logGeneratedAnswerHideAnswers = (): InsightAction =>
         action: 'hide',
         answer: {
           id: generativeQuestionAnsweringId!,
-          type: 'CRGA',
+          type: RGAType,
         },
       };
     },
   });
 
 export const logCopyGeneratedAnswer = (): InsightAction =>
-  makeInsightAnalyticsAction({
+  makeInsightAnalyticsActionFactory(
+    SearchPageEvents.generatedAnswerCopyToClipboard
+  )({
     prefix: 'analytics/generatedAnswer/copy',
     __legacy__getBuilder: (client, state) => {
       const generativeQuestionAnsweringId =
@@ -371,7 +389,7 @@ export const logCopyGeneratedAnswer = (): InsightAction =>
         action: 'copyToClipboard',
         answer: {
           id: generativeQuestionAnsweringId!,
-          type: 'CRGA',
+          type: RGAType,
         },
       };
     },