+++ packages/headless/src/features/generated-answer/generated-answer-slice.ts
@@ -24,6 +24,7 @@ import {
   setId,
   setAnswerMediaType,
   setRawAnswerMediaType,
+  setIsAnswerGenerated,
 } from './generated-answer-actions';
 import {getGeneratedAnswerInitialState} from './generated-answer-state';
 
@@ -131,5 +132,8 @@ export const generatedAnswerReducer = createReducer(
         state.fieldsToIncludeInCitations = [
           ...new Set(state.fieldsToIncludeInCitations.concat(action.payload)),
         ];
+      })
+      .addCase(setIsAnswerGenerated, (state, {payload}) => {
+        state.isAnswerGenerated = payload;
       })
 );